overlay33InitializeBuffers: close the addu, read the cluster off as1
The sixth word, the `addu` at +0xDC, was never a schedule tie. IDO
canonicalizes a commutative integer add so the heavier operand becomes the
left one, which is why `allocation + (width * height * 2)` and
`(width * height * 2) + allocation` both emit `addu t6,t5,v0`. An explicit
`(s32)` cast on the local makes it the heavier operand, the written order
survives, and the site becomes the target's `addu t6,v0,t5`. 75/81 -> 76/81.
The same rule, in its indexed-array form, closed the float pair in
overlay41AddSlot in this lane.

The remaining five words are now read off an as1 schedule trace rather than
inferred. `cc -Wa,-R` is byte-inert on .text at this TU's flags, and the
trace shows the whole cascade is one pick in a two-node basic block: as1
holds the allocation store and `beqz v0` as independent nodes, picks the
highest `aftercycles` and breaks ties on the lower ugen emission index. The
store's afternode list is empty (aftercycles 0); a branch is seeded at 1, so
it always wins and the store sinks into its delay slot, pushing `move` and
`li -64` down one slot each. The store already has the lower emission index,
so a tie would be enough -- and the trace shows the only thing that supplies
it: a second volatile reference in the block chains the store to it
(afternodes 1/1, aftercycles 2) and the target's order appears. Every such
reference costs an instruction, so the wall is the absence of a
zero-instruction successor, not an unreachable order.

Also newly falsified, all byte-flat: `volatile` on gOverlay33Allocation
itself (dropping it compiles identically, so it is not what pins this
schedule); every source-line layout of the store and the test, which rules
out a `.loc` boundary; `if (allocation)` and `if (!(allocation == 0))`;
casting the copy; testing `allocation` in the alignment test; and both
declaration orders of the allocation/original pair. Inverting the if/else
costs an instruction; driving the then-path from `allocation` costs nine
words; shifting `<< 1` and reassociating the product are flat.

verify OK 507341c0a40ca3e9a7cee969b396ee53facfb548; cleanroom and check-docs
pass.

+++ src/overlays/o033/overlay33InitializeBuffers.c
@@ -23,6 +23,46 @@ extern void overlay33AllocationFailedReloc(void);
  * line joins and the addition commutation are byte-flat; declaration changes
  * are flat/regressing, and all 25 relocation offsets/types align.
  *
+ * 2026-09-10, lane nm-ovlsmall: the sixth word, the `addu` at +0xDC, is
+ * closed. It was never a schedule tie: IDO canonicalizes a commutative
+ * integer add so that the heavier operand becomes the left one, which is why
+ * `allocation + (width * height * 2)` and `(width * height * 2) + allocation`
+ * both emit `addu t6,t5,v0`. An explicit `(s32)` cast on the local makes it
+ * the heavier operand, the written order survives, and the site becomes
+ * `addu t6,v0,t5`. (The same rule closed the float pair in
+ * overlay41AddSlot, where a struct-member reference replaced an indexed
+ * array reference.) Shifting `<< 1`, reassociating the product, and
+ * commuting in source are all byte-flat; `width * (height * 2)` regresses.
+ * The residual is now the five-word delay-slot cluster alone, and the
+ * following are newly falsified against it, all byte-flat: every source-line
+ * layout of the store and the test (same line, blank lines between,
+ * expressions split across lines -- so this is not a `.loc` boundary),
+ * `if (allocation)` and `if (!(allocation == 0))`, casting the copy,
+ * testing `allocation` instead of `original` in the alignment test, and both
+ * declaration orders of the pair. Inverting the if/else costs an
+ * instruction (82); driving the then-path from `allocation` costs nine more
+ * words.
+ *
+ * The five-word cluster is now read directly off an as1 schedule trace
+ * (`cc -Wa,-R`, byte-inert on .text here: the traced object's `.text` sha1 is
+ * the untraced one's). The whole cascade is ONE pick in a two-node basic
+ * block. as1 holds `sw v0,0(s0)` (the allocation store) and `beqz v0` in a
+ * block of exactly two independent nodes, and picks by highest `aftercycles`,
+ * breaking ties on the lower ugen emission index. The store's afternode list
+ * is empty, so its aftercycles is 0; a branch is seeded at 1. The branch
+ * therefore always wins, the store sinks into its delay slot, and the `move`
+ * and `li -64` each slide down one slot behind it. The store has the LOWER
+ * emission index, so a mere tie would be enough -- aftercycles 1 would do it.
+ * Only a dependence successor can supply that, and the trace shows exactly
+ * how: adding any second volatile reference to the block chains the store to
+ * it (`afternodes: 1/1`, aftercycles 2), the store is picked first, and the
+ * target's order appears. Every such reference costs an instruction (82
+ * against the target's 81), so the residual is not that the order is
+ * unreachable but that no zero-instruction source form gives that store a
+ * successor. Newly measured: `volatile` on gOverlay33Allocation is itself
+ * byte-flat here -- dropping the qualifier compiles to the identical object,
+ * so the barrier is not what pins the current schedule.
+ *
  * 2026-09-09, lane fin-misc: the cluster read out exactly. Five of the six
  * words are the same seven instructions in two orders. The target emits
  * `sw`, `beqz`, `move`(delay), `andi`, `beqz`, `li -64`(delay), `and`; the
@@ -63,7 +103,8 @@ void overlay33InitializeBuffers(void) {
                 allocation = original;
             }
             gOverlay33AlignedBuffers[0] = allocation;
-            gOverlay33AlignedBuffers[1] = allocation + (width * height * 2);
+            gOverlay33AlignedBuffers[1] =
+                (s32)allocation + (width * height * 2);
             gOverlay33ActiveBuffer =
                 gOverlay33AlignedBuffers[gOverlay33BufferIndex];
             if (overlay33InitializeBufferReloc(&gOverlay33InitializeContext,
@@ -84,10 +125,10 @@ void overlay33InitializeBuffers(void) {
 
 /* PLATEAU-HANDOFF:overlay33InitializeBuffers:start
  * symbol: overlay33InitializeBuffers
- * score: 75/81 words
+ * score: 76/81 words
  * frame: 0x38
  * relocations: 25
- * first-mismatch: +0x4C
- * summary: Six words are two as1 delay-slot fill choices plus the coupled addu; uopt sinks the partially dead `original` copy past the null test, leaving the store as the only fill above the branch.
+ * first-mismatch: +0x74
+ * summary: Five words are two as1 delay-slot fill choices; uopt sinks the partially dead `original` copy past the null test, leaving the store as the only fill above the branch. The coupled addu is closed by an operand-weight cast.
  * PLATEAU-HANDOFF:overlay33InitializeBuffers:end
  */
